+++ forui/test/src/widgets/avatar/avatar_golden_test.dart
@@ -8,7 +8,7 @@ import 'package:material_ui/material_ui.dart';
 
 import 'package:forui/forui.dart';
 
-import '../test_scaffold.dart';
+import '../../test_scaffold.dart';
 
 void main() {
   for (final theme in TestScaffold.themes) {